successful use of flyway and finish question comment

+++ src/main/resources/db/migration/V2__create_question_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE question
 (
-    id int AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     title varchar(50),
     description text,
     gmt_create BIGINT,
